Put To JSON Functions in Separate Class
