+++ vcell-core/src/main/java/cbit/vcell/resource/ErrorUtils.java
@@ -1,10 +1,7 @@
 package cbit.vcell.resource;
 
-import java.io.IOException;
-import java.text.SimpleDateFormat;
-import java.util.Calendar;
-
 import cbit.vcell.client.server.ClientServerInfo;
+import cbit.vcell.util.AmplistorUtils;
 import com.google.gson.Gson;
 import org.apache.http.client.methods.CloseableHttpResponse;
 import org.apache.http.client.methods.HttpPost;
@@ -20,7 +17,9 @@ import org.vcell.util.StackTraceUtils;
 import org.vcell.util.document.UserLoginInfo;
 import org.vcell.util.document.VCellSoftwareVersion;
 
-import cbit.vcell.util.AmplistorUtils;
+import java.io.IOException;
+import java.text.SimpleDateFormat;
+import java.util.Calendar;
 
 public class ErrorUtils {
